+++ pages/auth/manage/change-my-password.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import React from 'react';
 import { getSSRPropsDefaultAuth } from '../../../src/core/utils/commonGetSSRProps';
-import ChangePassword from '../../../src/stories/containers/AuthChange/ChangePassword/ChangePassword';
+import ChangePassword from '../../../src/stories/containers/Auth/ChangePassword/ChangePassword';
 import UserManagerLayout from '../../../src/stories/containers/Users/UsersManager/UserManagerLayout';
 import { ManagerTabs } from '../../../src/stories/containers/Users/UsersManager/managerTabsEnum';
 import lightTheme from '../../../styles/theme/light';
