+++ pages/login/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { featureFlags } from '../../feature-flags/feature-flags';
 import { CURRENT_ENVIRONMENT } from '../../src/config/endpoints';
-import Login from '../../src/stories/containers/AuthChange/Login/LoginContainer';
+import Login from '../../src/stories/containers/Auth/Login/LoginContainer';
 import type { NextPage } from 'next';
 
 const LoginPage: NextPage = () => <Login />;
