+++ src/stories/components/UserCard/UserCard.tsx
@@ -9,7 +9,7 @@ import { useThemeContext } from '../../../core/context/ThemeContext';
 import { ButtonType } from '../../../core/enums/buttonTypeEnum';
 import { getColorRole } from '../../../core/utils/colors';
 import { capitalizeWordWithoutConvertLowerCase, getCorrectRoleApi } from '../../../core/utils/string';
-import { ENABLE_DISABLE_USER_REQUEST } from '../../containers/AuthChange/ToggleAccountsActiveFlag/ToggleAccountsActiveFlagAPI';
+import { ENABLE_DISABLE_USER_REQUEST } from '../../containers/Auth/ToggleAccountsActiveFlag/ToggleAccountsActiveFlagAPI';
 import ControlledSwitches from '../Button/ControlledSwitches/ControlledSwitches';
 import { CustomButton } from '../CustomButton/CustomButton';
 import AvatarPlaceholder from '../svg/avatar-placeholder';
